dsp: move iir generic math functions to top-level module scope
